removed float
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@4866 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ src/player_manager.cpp
@@ -404,10 +404,10 @@ TRY {
 		
 			//GET_CONFIG_VALUE("multiplayer.ping-interpolation-multiplier", int, pw, 3);
 			//slot.trip_time = (pw * ping + slot.trip_time) / (pw + 1);
-			float p = slot.net_stats.updatePing(ping);
+			int p = (int)slot.net_stats.updatePing(ping);
 			
-			slot.net_stats.updateDelta(delta1 + (int)p);
-			slot.net_stats.updateDelta(delta2 - (int)p);
+			slot.net_stats.updateDelta(delta1 + p);
+			slot.net_stats.updateDelta(delta2 - p);
 			LOG_DEBUG(("player %u: ping: %g ms, delta: %+d", (unsigned)id, slot.net_stats.getPing(), slot.net_stats.getDelta()));		
 		}
 		break;
